separate viewer modules from img_analyzer
